# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "atrangi HR ML Service",
    "version": "1.0.0"
  },
  "paths": {
    "/health/live": {
      "get": {
        "summary": "Live",
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Live Health Live Get"
                }
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "summary": "Ready",
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Ready Health Ready Get"
                }
              }
            }
          }
        }
      }
    },
    "/ml/parse-resume": {
      "post": {
        "summary": "Parse Resume",
        "description": "Accept multipart field `file` and return a validated ResumeSchema.",
        "operationId": "parse_resume_ml_parse_resume_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_parse_resume_ml_parse_resume_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeSchema"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ml/generate-embedding": {
      "post": {
        "summary": "Generate Embedding",
        "description": "Return a normalized 384-dimensional vector for a short passage.",
        "operationId": "generate_embedding_ml_generate_embedding_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbeddingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Request Entity Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_parse_resume_ml_parse_resume_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Unencrypted text-bearing PDF"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_parse_resume_ml_parse_resume_post"
      },
      "EducationItem": {
        "properties": {
          "degree": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Degree"
          },
          "institution": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution"
          },
          "graduation_year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2200.0,
                "minimum": 1900.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Graduation Year"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "degree",
          "institution",
          "graduation_year"
        ],
        "title": "EducationItem",
        "description": "A documented qualification; unknown details are represented as null."
      },
      "EmbeddingRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "EmbeddingRequest",
        "description": "A sentence or short paragraph; overlong token sequences are rejected."
      },
      "EmbeddingResponse": {
        "properties": {
          "embedding": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 384,
            "minItems": 384,
            "title": "Embedding"
          },
          "dimensions": {
            "type": "integer",
            "const": 384,
            "title": "Dimensions",
            "default": 384
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "embedding"
        ],
        "title": "EmbeddingResponse",
        "description": "A normalized, finite 384-dimensional vector for pgvector vector(384)."
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Stable machine-readable failure code and safe human-readable message."
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Consistent failure envelope; never includes resume text or model output."
      },
      "ResumeSchema": {
        "properties": {
          "candidate_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 320
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "skills": {
            "items": {
              "type": "string",
              "maxLength": 150,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 200,
            "title": "Skills"
          },
          "experience_years": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Experience Years"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationItem"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Education"
          },
          "summary": {
            "type": "string",
            "maxLength": 3000,
            "title": "Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "candidate_name",
          "email",
          "skills",
          "experience_years",
          "education",
          "summary"
        ],
        "title": "ResumeSchema",
        "description": "Extracted facts. All keys are required; nullable fields may be null."
      }
    }
  }
}